refactor: write errors in middleware instead of doing manually (#98)
This PR continues my effort in reducing duplicate code, stop using `deps` just for the logger (and eventually we will be able to get rid of it completely).

Of course it's a quite big PR because there's a lot of duplication and I did a bunch of "search and replace", but 🤞 we are improving the situation little by little.