feat: add Sentry integration (#119)
This PR adds an initial Sentry integration by using the `ginsentry` package contained in emeris-utils. When a DSN is configured we can start monitoring performances for each endpoint in the Sentry dashboard.

In the future we are going to add more "spans" so we can see a more detailed view about where the time is spent (i.e. db access).