feat: put new /chains endpoint behind feature flag (#128)
This PR restores the original `/chains` endpoint we had before #118, since this requires a cockroachdb version that currently we don't have in staging.

We can now proceed with api-server deployments while we wait for cockroach to be upgraded in staging, and not remain stuck 😁 

I introduced a simple `fflag` package to manage feature flags, it's the simplest thing I could think of but I think it's sufficient for now. We can move into other libraries or even microservice if we want to do fancy things in the future.