refac: use only private mocks for ported methods
Only `usecase` and `api/chains` are impacted.

The problem is it doesn't compile because one dependency returns an
other dependency (SDKServiceClients which returns SDKServiceClient).
This makes the injection of the dependency complicated because from the
compiler point of view, they do not return the same types (even if in
term of interfaces, they expose the same set of methods).

One way to fix this is to not inject SDKServiceClients in usecase, but
the SDKServiceClient (without s) only. That means we can no longer call
NewApp at the initialization phase, but for each request because the
SDKServiceClient injected will be different according to the chain
version. This isn't a big problem though, but I tend to prefer to have a
single usecase.App.

If you see an other solution, I'm all ears.